backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
requests
beautifulsoup4
pyjwt
email-validator
python-multipart
passlib[bcrypt]
bcrypt
